Hack to force inheritance of fmtp for AMR/AMR-WB via 'rtp_force_audio_fmtp' when inheriting codec
This fixes an issue where incompatible mode-set is inserted in the 200 OK back to the originator in a bridged call using 'inherit_codec'.

Without this fix the 'default-bitrate' setting from the AMR/AMR-WB configuration will always be inserted in the 200 OK back to the originator if the originator's initial INVITE does not contain any mode-set.