Fix an issue where audio stream gets stuck as inactive after re-INVITE
If a channel has established media streams with FS and then gets bridged to an external part which answers with a held (a=inactive) audio stream, the A side will also be marked as inactive.
If B then re-INVITEs and un-holds the audio (a=sendrecv), this 'sdp_type' check would prevent the A side from getting updated with the expected sendrecv value.